refactor the language part
