+++ src/app/[variants]/(main)/settings/common/features/Theme/index.tsx
@@ -4,7 +4,6 @@ import { Form, type ItemGroup, SelectWithImg, SliderWithInput } from '@lobehub/u
 import { Select } from 'antd';
 import isEqual from 'fast-deep-equal';
 import { Monitor, Moon, Sun } from 'lucide-react';
-import { useRouter } from 'next/navigation';
 import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -22,7 +21,7 @@ type SettingItemGroup = ItemGroup;
 
 const Theme = memo(() => {
   const { t } = useTranslation('setting');
-  const router = useRouter();
+
   const [form] = Form.useForm();
   const settings = useUserStore(settingsSelectors.currentSettings, isEqual);
   const themeMode = useUserStore(userGeneralSettingsSelectors.currentThemeMode);
@@ -33,7 +32,6 @@ const Theme = memo(() => {
 
   const handleLangChange = (value: Locales) => {
     switchLocale(value);
-    router.refresh();
   };
 
   const theme: SettingItemGroup = {
