+++ src/chains/__tests__/summaryAgentName.test.ts
@@ -1,6 +1,6 @@
 import { Mock, describe, expect, it } from 'vitest';
 
-import { globalHelpers } from '@/store/user/helpers';
+import { globalHelpers } from '@/store/global/helpers';
 
 import { chainSummaryAgentName } from '../summaryAgentName';
