+++ src/chains/__tests__/summaryDescription.test.ts
@@ -1,6 +1,6 @@
 import { Mock, describe, expect, it, vi } from 'vitest';
 
-import { globalHelpers } from '@/store/user/helpers';
+import { globalHelpers } from '@/store/global/helpers';
 
 import { chainSummaryDescription } from '../summaryDescription';
