+++ src/chains/__tests__/summaryTags.test.ts
@@ -1,6 +1,6 @@
 import { Mock, describe, expect, it } from 'vitest';
 
-import { globalHelpers } from '@/store/user/helpers';
+import { globalHelpers } from '@/store/global/helpers';
 
 import { chainSummaryTags } from '../summaryTags';
