+++ src/const/settings/common.ts
@@ -2,6 +2,5 @@ import { UserGeneralConfig } from '@/types/user/settings';
 
 export const DEFAULT_COMMON_SETTINGS: UserGeneralConfig = {
   fontSize: 14,
-  language: 'auto',
   themeMode: 'auto',
 };
