+++ src/store/global/selectors/general.ts
@@ -1,9 +1,25 @@
-import { GlobalStore } from '@/store/global';
+import { DEFAULT_LANG } from '@/const/locale';
+import { Locales } from '@/locales/resources';
+import { isOnServerSide } from '@/utils/env';
 
+import { GlobalState } from '../initialState';
 import { systemStatus } from './systemStatus';
 
-const language = (s: GlobalStore) => systemStatus(s).language || 'auto';
+const language = (s: GlobalState) => systemStatus(s).language || 'auto';
+
+const currentLanguage = (s: GlobalState) => {
+  const locale = language(s);
+
+  if (locale === 'auto') {
+    if (isOnServerSide) return DEFAULT_LANG;
+
+    return navigator.language as Locales;
+  }
+
+  return locale as Locales;
+};
 
 export const globalGeneralSelectors = {
+  currentLanguage,
   language,
 };
