--- src/store/user/helpers.ts
@@ -1,9 +1,0 @@
-import { userGeneralSettingsSelectors } from './slices/settings/selectors';
-import { useUserStore } from './store';
-
-const getCurrentLanguage = () =>
-  userGeneralSettingsSelectors.currentLanguage(useUserStore.getState());
-
-export const globalHelpers = {
-  getCurrentLanguage,
-};
